refactor(mountains): extract snow-surface helpers into mountains/snow-surface.ts
Move the procedural snow look — createSnowAlbedoTexture / createSnowNormalTexture
(tileable CanvasTextures) and the applySnowVertexColors / applySmoothShadingNormals
geometry passes — into src/mountains/snow-surface.ts. Render-only (never touches the
height field), so physics is unaffected; the forest treeline tint keeps reading the
shared forestDensityField. createTerrain imports them from the new module.

Verified: tsc + lint clean, terrain (9/9), regression (10/10), verify 49/49.
